requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
psycopg[binary]==3.2.9
sqlalchemy>=2.0.25
python-dotenv==1.0.0
alembic==1.13.1